backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg[binary]==3.2.3
pydantic==2.9.2
python-dotenv==1.0.1
openai==1.51.2
numpy==1.26.4
